refactor: Move rendering into common static class
